+++ modules/actions/docker-utils.js
@@ -1,0 +1,8 @@
+function dockerHostUserParams(libdragonInfo) {
+  const { uid, gid } = libdragonInfo.userInfo;
+  return ['-u', `${uid >= 0 ? uid : ''}:${gid >= 0 ? gid : ''}`];
+}
+
+module.exports = {
+  dockerHostUserParams,
+};