+++ modules/actions/stop.js
@@ -18,4 +18,11 @@ module.exports = {
   name: 'stop',
   fn: stop,
   showStatus: false, // This will only print out the id
+  usage: {
+    name: 'stop',
+    summary: 'Stop the container for current project.',
+    description: `Stop the container assigned to the current libdragon project.
+
+      Must be run in an initialized libdragon project.`,
+  },
 };