+++ modules/actions/utils.js
@@ -19,12 +19,8 @@ const {
   ParameterError,
 } = require('../helpers');
 
+const { dockerHostUserParams } = require('./docker-utils');
 const { installNPMDependencies } = require('./npm-utils');
-
-function dockerHostUserParams(libdragonInfo) {
-  const { uid, gid } = libdragonInfo.userInfo;
-  return ['-u', `${uid >= 0 ? uid : ''}:${gid >= 0 ? gid : ''}`];
-}
 
 const installDependencies = async (libdragonInfo) => {
   const buildScriptPath = path.join(
@@ -219,7 +215,6 @@ module.exports = {
   destroyContainer,
   checkContainerRunning,
   checkContainerAndClean,
-  dockerHostUserParams,
   tryCacheContainerId,
   runGitMaybeHost,
   mustHaveProject,